feat: Add AutoMapper to UsersControllers
This commit adds the AutoMapper library to the UsersControllers class in order to map the AppUser entity to the MemberDto DTO. It also updates the GetUsers and GetUser methods to return MemberDto objects instead of AppUser objects.

Note: This commit message follows the established conventions in the repository.

Listo seccion 8, hasta la parte de 96